+++ cercle-5/miniRT/includes/functions/primitives/scene.h
@@ -6,7 +6,7 @@
 /*   By: abdoali <abdoali@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/11 11:45:00 by abdoali           #+#    #+#             */
-/*   Updated: 2026/04/01 14:01:07 by abdoali          ###   ########.fr       */
+/*   Updated: 2026/04/03 11:59:48 by abdoali          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -75,7 +75,7 @@ t_scene					*parse_file(const char *path, void *mlx);
 
 /* Internal helpers */
 int						scene_material_allocate_slot(t_scene *scene);
-int						scene_allocate_object_slot(void **ptr, int *count,
-							int *cap, size_t sz);
+int						scene_allocate_object_slot(void **ptr, size_t *count,
+							size_t *cap, size_t sz);
 
 #endif
