+++ cercle-5/miniRT/packages/libft/srcs/data/stack/ft_stack_reserve.c
@@ -6,7 +6,7 @@
 /*   By: abdoali <abdoali@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/25 14:22:05 by abdoali           #+#    #+#             */
-/*   Updated: 2026/03/25 14:29:02 by abdoali          ###   ########.fr       */
+/*   Updated: 2026/04/03 11:49:38 by abdoali          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ int	ft_stack_reserve(t_stack *s, size_t additional)
 	if (!s)
 		return (-1);
 	need = s->size + additional;
-	if (need <= (size_t)s->cap)
+	if (need <= s->cap)
 		return (0);
 	newcap = s->cap;
 	if (newcap == 0)
