+++ cercle-5/miniRT/srcs/gui/editor/crud/delete/simple.c
@@ -17,13 +17,13 @@ void	delete_sel_sp_pl(t_scene *sc, t_selection *sel)
 	int	i;
 
 	i = sel->index;
-	if (sel->type == TYPE_SPHERE && i < sc->sphere_count)
+	if (sel->type == TYPE_SPHERE && (size_t)i < sc->sphere_count)
 	{
 		ft_memmove(sc->spheres + i, sc->spheres + i + 1,
 			(sc->sphere_count - i - 1) * sizeof(t_sphere));
 		sc->sphere_count--;
 	}
-	else if (sel->type == TYPE_PLANE && i < sc->plane_count)
+	else if (sel->type == TYPE_PLANE && (size_t)i < sc->plane_count)
 	{
 		ft_memmove(sc->planes + i, sc->planes + i + 1,
 			(sc->plane_count - i - 1) * sizeof(t_plane));
